linjing/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
faiss-cpu==1.10.0
frozenlist==1.5.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
loguru==0.7.3
multidict==6.4.3
numpy==2.2.4
openai==1.75.0
packaging==24.2
pillow==11.2.1
portalocker==2.10.1
propcache==0.3.1
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
PyYAML==6.0.2
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.0
